fix connectors tests
