+++ core/src/test/scala/com/raphtory/BaseRaphtoryAlgoTest.scala
