+++ core/src/test/scala/com/raphtory/TestUtils.scala
@@ -14,6 +14,8 @@ import java.nio.file.Paths
 import scala.sys.process._
 
 object TestUtils {
+  private val logger: Logger = Logger(LoggerFactory.getLogger(this.getClass))
+
   def manageTestFile(fileDefinition: Option[(String, URL)]): Resource[IO, Any] =
     fileDefinition match {
       case None           => Resource.eval(IO.unit)
@@ -28,4 +30,33 @@ object TestUtils {
           }
         )
     }
+
+  def generateTestHash(outputDirectory: String, jobId: String): String = {
+    val results = getResults(outputDirectory, jobId)
+    val hash = resultsHash(results)
+    logger.info(s"Generated hash code: '$hash'.")
+
+    hash
+  }
+
+  def resultsHash(results: IterableOnce[String]): String =
+    Hashing
+      .sha256()
+      .hashString(results.iterator.toSeq.sorted.mkString, StandardCharsets.UTF_8)
+      .toString
+
+  def getResults(outputDirectory: String, jobID: String): Iterator[String] = {
+    val files = new File(outputDirectory + "/" + jobID)
+      .listFiles()
+      .filter(_.isFile)
+
+    files.iterator.flatMap { file =>
+      val source = scala.io.Source.fromFile(file)
+      try source.getLines().toList
+      catch {
+        case e: Exception => throw e
+      }
+      finally source.close()
+    }
+  }
 }
